feat!: init WASM during decode, no top-level init
Separate the WebAssembly-based image decoder into a separate package
named @webtoon/psd-decoder. This is built separately using Vite, and
exports the `init` promise, as well as methods generated by wasm-pack.
This effectively "moves" the top-level `init` from @webtoon/psd to
@webtoon/psd-decoder.

The `generateRgba()` function in @webtoon/psd imports
@webtoon/psd-decoder and awaits the `init` promise, ensuring that the
WebAssembly module is ready before using any of its methods. This allows
the rest of @webtoon/psd to be treated like a regular JavaScript module,
and no longer exports an `init` promise.

The overall change means that the "asynchrony" of WebAssembly is safely
contained in `generateRgba()`, instead of spread all over the library.
On the flip side, `generateRgba()` is now asynchronous, even though it
technically doesn't have to be.