feat!: export Group, Layer, Slice as types only
We only want these to exist as types that users can refer to.
We don't want users to directly instantiate these classes.

This change breaks the original `traverseNode()` example, which was
using `instanceof` to narrow a `Node` into `Psd`, `Layer`, `Group`.
From now on, users now have to check the `type` field to do so.
To make `traverseNode()` work, we replaced `Node` with a union type of
`Psd`, `Layer`, and `Group`. We also added `NodeParent` and `NodeChild`
types to better identify which types can be parents and which can be
children. The original `Node` interface has been renamed to `NodeBase`.